Extract out the shared constants from ObjC
