+++ Sources/Encrypted Logs/Extensions/EventLoggingFileUploadError.swift
@@ -1,5 +1,9 @@
 import Foundation
 
+#if SWIFT_PACKAGE
+import AutomatticTracksConstantsObjC
+#endif
+
 public enum EventLoggingFileUploadError: Error, LocalizedError {
     /// HTTP Errors should receive the error code, a localized error string, and the HTTP status code
     case httpError(String, String, Int)
@@ -17,6 +21,28 @@ public enum EventLoggingFileUploadError: Error, LocalizedError {
                 "System Cancelled Upload",
                 comment: "A message indicating that a file queued for upload was cancelled by the system"
             )
+        }
+    }
+}
+
+extension EventLoggingFileUploadError: CustomNSError {
+
+    public static var errorDomain: String {
+        return TracksErrorDomain
+    }
+
+    public var errorCode: Int {
+        switch self {
+            case .httpError(_, _, let statusCode): return statusCode
+            case .fileMissing: return TracksErrorCode.fileMissing.rawValue
+            case .cancelledByDelegate: return TracksErrorCode.operationCancelled.rawValue
+        }
+    }
+
+    public var failureReason: String? {
+        switch self {
+            case .httpError(_, let message, _): return message
+            default: return nil
         }
     }
 }
