+++ Sources/Model/ObjC/Common/TracksDeviceInformation.h
