+++ Sources/Model/ObjC/Common/TracksDeviceInformation.m
