+++ Sources/Model/ObjC/Common/UIApplication+Extensions.h
