+++ Sources/Model/ObjC/Common/UIApplication+Extensions.m
