+++ Sources/Model/ObjC/Common/WatchSessionManager.h
