+++ Sources/Model/ObjC/Common/WatchSessionManager.m
